pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Gabriel Brotas", email = "gabriel.brotas@wexinc.com"},
]
dependencies = [
    "fastapi>=0.78.0",
    "uvicorn>=0.17.6",
    "kafka-python>=2.0.2",
    "pydantic>=1.9.1",
    "fastavro>=1.5.0",
    "confluent-kafka>=1.8.2",
    "requests>=2.28.0",
    "avro-python3>=1.10.2",
    "python-dotenv>=0.20.0"]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]

[tool.pdm.scripts]
"producer" = "uvicorn src.producer.main:app --reload"
"consumer" = "python3 src/consumer/main.py"
"kafka:up" = "docker-compose up -d --build"

"aiven:p" = "python3 src/aiven/producer.py"
"aiven:pa" = "python3 src/aiven/producer-avro.py"
"aiven:c" = "python3 src/aiven/consumer.py"
"aiven:ca" = "python3 src/aiven/consumer-avro.py"

"confluent:p" = "python3 src/confluent/producer.py"
"confluent:pa" = "python3 src/confluent/producer-avro.py"
"confluent:c" = "python3 src/confluent/consumer.py"
"confluent:ca" = "python3 src/confluent/consumer-avro.py"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
